Merge branch 'master' of https://github.com/koheiw/Newsmap

# Conflicts:
#	DESCRIPTION
#	NAMESPACE
#	man/check_name.Rd
#	man/construct.Rd
#	man/findNames.Rd
#	man/findNames2.Rd
#	man/findNames3.Rd
#	man/findNames4.Rd
#	man/getCasedTypes.Rd
#	man/getKeywords.Rd
#	man/gscore.Rd
#	man/joinNames.Rd
#	man/lexicon_en.Rd
#	man/lexicon_ru.Rd
#	man/makeGeoDictionary.Rd
#	man/makeGeoDictionary2.Rd
#	man/readLexicon.Rd
#	man/removePadding.Rd
#	man/removeShortFeatures.Rd
#	man/removeSpecialFeatures.Rd
#	man/selectCasedFeatures.Rd
#	man/selectNames.Rd
#	man/stemNames.Rd

+++ R/construct.R
@@ -5,27 +5,31 @@
 #' @param smooth smoother for coditional frequency
 #' @param sep_keyword separator for keywords in lexicon
 #' @export
-make_dictionary <- function(toks, dict, concatenator = ' ', weight = 1, smooth = 10e-5){
+construct <- function(toks, dict, concatenator = ' ', weight = 1, smooth = 1, min_count_seq = 2){
 
     if(!is.dictionary(dict)) stop("Dictionary has to be a quanteda dictionary object")
     if(length(dict) == 0) stop("Dictionary has not entry")
 
     # Dicitonary lookup
-    cat("Searching for dictionary words\n")
+    cat("Searching for dictionary words...\n")
     keys <- quanteda::tokens_lookup(toks, dict, levels = 3)
     mx_key <- quanteda::dfm(keys, tolower = FALSE)
 
     # Feature selection
-    file_seqs <- 'sequences_capital.RDS'
+    md5 <- digest::digest(toks , 'md5')
+    file_seqs <- paste0('.cache_', min_count_seq, '_', md5 ,'.RDS', collapse = '')
+    cat("Identifying multi-word names...\n")
+    if (file.exists(file_seqs)) {
+        cat("Loading cached multi-word names:", file_seqs, "\n")
+        seqs <- readRDS(file=file_seqs)
+    } else {
+        seqs <- quanteda::sequences(toks, '^[A-Z]', valuetype = 'regex', case_insensitive = FALSE,
+                                    min_count = min_count_seq)
+        saveRDS(seqs, file=file_seqs)
+    }
 
-    # if (file.exists(file_seqs)) {
-    #   cat("Loading cached multi-word names:", file_seqs, "\n")
-    #   seqs <- readRDS(file=file_seqs)
-    # } else {
-    #   cat("Idenityfing multi-word names\n")
-    seqs <- quanteda::sequences(toks, '^[A-Z]', valuetype = 'regex', case_insensitive = FALSE, min_count = 2)
-    saveRDS(seqs, file=file_seqs)
-    # }
+    cat("Concatenating multi-word names...\n")
+    seqs <- seqs[seqs$p < 0.001,]
     toks <- quanteda::tokens_compound(toks, seqs$sequence, valuetype = 'fixed', concatenator = ' ')
     toks <- quanteda::tokens_select(toks, '^[A-Z]', valuetype = 'regex', padding = TRUE, case_insensitive = FALSE)
     mx <- dfm(toks)
@@ -36,48 +40,38 @@ make_dictionary <- function(toks, dict, concatenator = ' ', weight = 1, smooth =
     colnames(mx_dic) <- colnames(mx)
     rownames(mx_dic) <- colnames(mx_key)
 
+    cat("Scoring words for countries: ")
     for(key in sort(featnames(mx_key))){
 
-        cat(key, "\n")
+        cat(key, " ", sep = "")
         rownames(mx) <- ifelse(as.vector(mx_key[,key]) > 0, 'T', 'R')
         mx2 <- quanteda::dfm_compress(mx, margin = 'documents')
 
         missing <- setdiff(c('T', 'R'), rownames(mx2))
         attr(mx2, 'Dim')[1L] <- attr(mx2, 'Dim')[1L] + length(missing)
         attr(mx2, 'Dimnames')$docs <- c(attr(mx2, 'Dimnames')$docs, missing)
-        #flag_zero <- as.vector(mx2['T',] == 0) # flag words do not occur
 
         # Scoring words
-
-        #sum <- sum(mx2) + 1
-        mx2 <- mx2 + 1
+        mx2 <- mx2 + smooth
         sums <- Matrix::rowSums(mx2)
-
         if (sums['T'] > 1) {
-            #mx3 <- (mx2 ^ weight) / sums # conditional likelihood
-            #if (missing(smooth)) smooth <- sums['T'] / sum #
-            #cat("sums\n")
-            #print(sums)
-            #print(smooth)
-            #mx3 <- ((mx2 ^ weight) + smooth) / (sums + smooth) # conditional likelihood
             mx3 <- (mx2 ^ weight) / (sums) # conditional likelihood
-            print(head(mx3, 20))
-            lr <- as.vector(log(mx3['T',]) - log(mx3['R',])) # # calculate likelihood-ratio
-            #lr[flag_zero] <- 0 # fill with zero if words do no occur
+            lr <- as.vector(log(mx3['T',]) - log(mx3['R',])) # calculate likelihood-ratio
             mx_dic[key,] <- lr
         } else {
             mx_dic[key,] <- NULL
         }
     }
+    cat("\n")
     mx_dict_sps <- as(mx_dic, "sparseMatrix")
     return(mx_dict_sps)
 }
 
 #' @export
-topEntries <-function(dict, code, n=20){
-  ents <- dict[code,]
-  ents <- ents[order(ents, decreasing = TRUE)]
-  print(head(ents, n))
-  cat(sum(ents > 0), 'non-zero entries\n')
+top_entries <-function(dict, code, n=20){
+    ents <- dict[code,]
+    ents <- ents[order(ents, decreasing = TRUE)]
+    print(head(ents, n))
+    cat(sum(ents > 0), 'non-zero entries\n')
 }
